+++ docs/changelog.rst
@@ -8,7 +8,8 @@ Fixed Bugs
 ******************
 
 - docs/_build is excluded from source tarball builds
-
+- use table.append_column() instead of column._set_parent() in
+  ChangesetColumn.add_to_table()
 
 0.7 (2011-05-27)
 ---------------------------